resolve: port #9705 merge-time constraint gate into BranchMergeOrchestrator
develop replaced BranchMerger with GraphMerger/BranchMergeOrchestrator; stable's
#9705 added MergeConstraintValidator + an unresolved-conflict gate inside the
now-deleted BranchMerger.merge(). Re-home both gates into GraphMerger.merge(),
where the diff is freshened and the diff lock is held: run them before the graph
is mutated, raising MergeConstraintsViolatedError / MergeConflictsUnresolvedError
(the latter replacing the prior plain ValidationError, same wording). Wire
MergeConstraintValidator and MergeSchemaAnalyzer into GraphMerger via the builder.
